* Light unfixed
git-svn-id: https://rendergirl.googlecode.com/svn/trunk@54 636cc87a-7a7b-1416-3ae0-be5257d649da

+++ Core/CLMath.h
@@ -49,7 +49,7 @@ cl_float3 subtract(cl_float3& a, cl_float3& b)
 	return result;
 }
 
-inline double length(cl_float3& v)
+inline float length(cl_float3& v)
 {
 	return sqrt((v.s[0] * v.s[0]) + (v.s[1] * v.s[1]) + (v.s[2] * v.s[2]));
 }
